Implementation of Psi references and search
The previous mechanism was not correct in the sense that a macro was
not referenced by its method. @Macro("Foo#bar") will contribute two
references now. One for "Foo" and one for its method "bar".

The reference search is also correct in the sense that we don't use
IntelliJ's default scope (module with dependencies) but build our own.
The macro inside a web project might as well be referenced from generic
code which doesn't share a dependency in IntelliJ's cosmos.