Linkage of Psi references
The Psi references were not linked correct in the sense that
@DelegateWeb would reference classes for the Android scope since the
resolve scope was allProject instead of the delegate search scope of
the platform. However even this scope is too wide (e.g. DelegateWeb
in a generic project) and therefore a new target platform is
introduced when building a search scope.