Implement Caching of DefracConfig
The old cache was based on IntelliJ IDEAs system for PsiElement keys
which is not what we want since the DefracConfig is bound to a
VirtualFile. There might be a better API but since we can use Guava there
isn't much OpenAPI could offer us which would make this more pleasant.